Full npm workflow.
- moved the overview to the top
- HTML driven template
- Webpack bundling
- removed bower
