--- bower.json
@@ -1,31 +1,0 @@
-{
-  "name": "waveform.prototyping.bbc.co.uk",
-  "version": "0.1.0",
-  "authors": [
-    "BBC R&D"
-  ],
-  "description": "Browser-based audio waveform visualisation Webpage",
-  "main": "index.html",
-  "keywords": [
-    "peaks",
-    "bbc",
-    "waveform"
-  ],
-  "license": "Apache-2",
-  "homepage": "http://waveform.prototyping.bbc.co.uk",
-  "private": true,
-  "ignore": [
-    "**/.*",
-    "node_modules",
-    "bower_components",
-    "test",
-    "tests"
-  ],
-  "dependencies": {
-    "KineticJS": "http://d3lp1msu2r81bx.cloudfront.net/kjs/js/lib/kinetic-v4.5.5.min.js",
-    "eventEmitter": "~4.2.3",
-    "peaks.js": "master",
-    "requirejs": "~2.1.14"
-  },
-  "devDependencies": {}
-}
