+++ index.html
@@ -85,7 +85,12 @@
       <div class="grid-row clearfix" id="waveform">
         <div class="grid-col grid-12">
           <h2>Example Demo</h2>
-          <div id="peaks-container"></div>
+          <div id="peaks-container">
+            <div class="waveform">
+              <div class="overview-container"></div>
+              <div class="zoom-container"></div>
+            </div>
+          </div>
           <div class="peaks-control">
             <div class="peaks-audio">
               <audio id="peaks-audio" controls>
@@ -133,14 +138,11 @@
 
     <footer class="masthead">
       <div class="grid">
-        <span>&copy; Copyright 2014 BBC Research and Development</span>
+        <span>&copy; Copyright 2015 BBC Research and Development</span>
         <span class="audio_licence"><img src="cc-logo.png" alt="" height="36">The audio in this demo is taken from the <a href="http://en.wikipedia.org/wiki/File:Alice_walker_bbc_radio4_desert_island_discs_19_05_2013.flac">radio show Desert Island</a>, used under the terms of the <a href="http://creativecommons.org/licenses/by/3.0/">Creative Commons 3.0 Unported license</a>.</span>
       </div>
     </footer>
 
-
-    <script src="bower_components/KineticJS/index.js" defer></script>
-    <script src="bower_components/requirejs/require.js" defer></script>
-    <script src="main.js" defer></script>
+    <script src="bundle.js" async></script>
 </body>
 </html>
